Rate type data type issue

See merge request mislbd-ng/services-ng/ababil-foreign-remittance!142

+++ foreign-remittance-impl/src/main/java/com/mislbd/ababil/foreignremittance/repository/schema/RemittanceTransactionEntity.java
@@ -97,7 +97,7 @@ public class RemittanceTransactionEntity extends BaseEntity {
 
   @Deprecated
   @Column(name = "OPERATING_RATE_TYPE")
-  private Long operatingRateTypeId;
+  private Integer operatingRateTypeId;
 
   @Deprecated
   @Column(name = "OPERATING_RATE")
